requirements.txt
psycopg2-binary
faker
dotenv
confluent-kafka[avro]